Merge develop into master

+++ jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/builders/OffsetB11Builder.java
@@ -8,7 +8,7 @@ public class OffsetB11Builder {
       throw new UnsupportedOperationException();
    }
 
-   public static int offsetB11(BigDecimal offset) {
-      return offset.scaleByPowerOfTen(2).intValue();
+   public static Long offsetB11(BigDecimal offset) {
+      return offset.scaleByPowerOfTen(2).longValue();
    }
 }
