+++ jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/builders/OffsetB12Builder.java
@@ -8,7 +8,7 @@ public class OffsetB12Builder {
       throw new UnsupportedOperationException();
    }
 
-   public static int offsetB12(BigDecimal offset) {
-      return offset.scaleByPowerOfTen(2).intValue();
+   public static Long offsetB12(BigDecimal offset) {
+      return offset.scaleByPowerOfTen(2).longValue();
    }
 }
