+++ jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/builders/OffsetB14Builder.java
@@ -8,7 +8,7 @@ public class OffsetB14Builder {
       throw new UnsupportedOperationException();
    }
 
-   public static int offsetB14(BigDecimal offset) {
-      return offset.scaleByPowerOfTen(2).intValue();
+   public static Long offsetB14(BigDecimal offset) {
+      return offset.scaleByPowerOfTen(2).longValue();
    }
 }
