+++ jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/builders/OffsetB16Builder.java
@@ -8,7 +8,7 @@ public class OffsetB16Builder {
       throw new UnsupportedOperationException();
    }
 
-   public static int offsetB16(BigDecimal offset) {
-      return offset.scaleByPowerOfTen(2).intValue();
+   public static Long offsetB16(BigDecimal offset) {
+      return offset.scaleByPowerOfTen(2).longValue();
    }
 }
